requirements.txt
# Deep Learning Core
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
imageio>=2.30.0
numpy>=1.21.0

# Web UI
streamlit>=1.20.0
streamlit-image-comparison>=0.0.4